[FALSE-POSITIVE] salesforce.gompels.com ref #470
Ref #470

**Domains or links**
salesforce.gompels.com

**More Information**
How did you discover your web site or domain was listed here?
1. Google (via virustotal.com)

**Have you requested removal from other sources?**
Not listed in openphish or phishtank
Reported the parent domain (gompels.com) to Bitdefender as non-malicious
As we aren't using and don't intend to use salesforce.gompels.com we haven't made efforts to request removal -- my main concern is the impact on the parent domain (gompels.com)

**Additional context**
Last year the subdomain was set-up with an nginx reverse proxy to the salesforce login page by one of our staff to test passing login from our SSO through, was never used but was not removed.

From the 28th Jan 2023 e-mail reports from Phishlabs reporting this subdomain as a malicious came through to a general enquiries e-mail which staff ignored as spam.
On the 9th of Feb 2023 it was raised to IT and the nginx vhost was taken down (the A record and certificates remained)
On the 23rd of Feb 2023 a large portion of the web services with subdomains under gompels.com were re-created (with new SSL certs issued) due to a cluster failure
On the 27th of Feb 2023 the main domain (gompels.com) was reported by a customer as being flagged by Bitdefender. virustotal reported 3 AV products flagged it. At this point the remaining A record for salesforce.gompels.com was removed
On the 28th of Feb 2023 an additional AV vendor flagged the main domain (gompels.com) as malicious and this github repo now lists salesforce.gompels.com in the `phishing-domains-NEW-today.txt` dated today. The lets-encrypt cert for salesforce.gompels.com was revoked

I am unsure how/why this domain has been listed as it hasn't proxied to salesforce for 20 days (at most would have displayed a default vhost nginx boilerplate page) and hasn't had a working A record for nearly 24 hours.
I don't know if the re-issuing of a large number of our subdomains has impacted the domain's reputation which is why I mention it.

Any help understanding why this subdomain has appeared on this list _after_ it's removal and any potential impact on the parent domain would be greatly appreciated